requirements.txt
Flask==2.3.2
flask-mysqldb==1.0.1
mysql-connector-python==8.3.0
Werkzeug==2.3.4
Jinja2==3.1.3
itsdangerous==2.1.2
MarkupSafe==2.1.5
WTForms==3.1.2
